refactor(api-infra): extract AWS API handler comp-root into @webiny/api-infra-aws
Phase 1 of the api-infra composition layer (see plans/api-infra-composition-layer.md).

The GraphQL handler's composition root — ~30 feature registrations in a specific
order — lived in the scaffolding template
`_templates/appTemplates/api/graphql/src/index.ts`, exercised only by full e2e
deploy. That's why wiring/fresh-install bugs (e.g. "No registration found for
FileModel") surface only in CI e2e, never in a test.

Move it verbatim (order preserved) into a real package, @webiny/api-infra-aws,
behind `createWebinyApiHandler(config)`. The only project-specific input is
`extensions`; everything else is standard AWS/env wiring owned by the package.
The template collapses to a 3-line shim. This makes the comp-root a real,
integration-testable unit (Phase 2 adds the fresh-install test).

No behavior change. Build + adio + lint green; ESM imports validated.
