+++ packages/api-infra-aws/src/index.ts
@@ -1,0 +1,2 @@
+export { createWebinyApiHandler } from "./createWebinyApiHandler.js";
+export type { CreateWebinyApiHandlerConfig } from "./createWebinyApiHandler.js";
