+++ packages/api-infra-aws/tsconfig.build.json
@@ -1,0 +1,117 @@
+{
+  "extends": "../../tsconfig.build.json",
+  "include": ["src"],
+  "references": [
+    { "path": "../api-aco/tsconfig.build.json" },
+    { "path": "../api-aco-ddb/tsconfig.build.json" },
+    { "path": "../api-audit-logs/tsconfig.build.json" },
+    { "path": "../api-audit-logs-ddb/tsconfig.build.json" },
+    { "path": "../api-core/tsconfig.build.json" },
+    { "path": "../api-core-ddb/tsconfig.build.json" },
+    { "path": "../api-file-manager/tsconfig.build.json" },
+    { "path": "../api-file-manager-aco/tsconfig.build.json" },
+    { "path": "../api-file-manager-s3/tsconfig.build.json" },
+    { "path": "../api-headless-cms/tsconfig.build.json" },
+    { "path": "../api-headless-cms-aco/tsconfig.build.json" },
+    { "path": "../api-headless-cms-ddb/tsconfig.build.json" },
+    { "path": "../api-headless-cms-scheduler/tsconfig.build.json" },
+    { "path": "../api-headless-cms-tasks/tsconfig.build.json" },
+    { "path": "../api-headless-cms-workflows/tsconfig.build.json" },
+    { "path": "../api-mailer/tsconfig.build.json" },
+    { "path": "../api-record-locking/tsconfig.build.json" },
+    { "path": "../api-scheduler/tsconfig.build.json" },
+    { "path": "../api-scheduler-aws/tsconfig.build.json" },
+    { "path": "../api-website-builder/tsconfig.build.json" },
+    { "path": "../api-website-builder-scheduler/tsconfig.build.json" },
+    { "path": "../api-website-builder-workflows/tsconfig.build.json" },
+    { "path": "../api-websockets/tsconfig.build.json" },
+    { "path": "../api-websockets-aws/tsconfig.build.json" },
+    { "path": "../api-websockets-ddb/tsconfig.build.json" },
+    { "path": "../api-workflows/tsconfig.build.json" },
+    { "path": "../aws-sdk/tsconfig.build.json" },
+    { "path": "../background-tasks/tsconfig.build.json" },
+    { "path": "../cognito/tsconfig.build.json" },
+    { "path": "../event-handler-aws/tsconfig.build.json" },
+    { "path": "../handler/tsconfig.build.json" },
+    { "path": "../handler-db/tsconfig.build.json" },
+    { "path": "../handler-graphql/tsconfig.build.json" },
+    { "path": "../webhooks/tsconfig.build.json" }
+  ],
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./dist",
+    "declarationDir": "./dist",
+    "paths": {
+      "~/*": ["./src/*"],
+      "~tests/*": ["./__tests__/*"],
+      "@webiny/api-aco/*": ["../api-aco/src/*"],
+      "@webiny/api-aco": ["../api-aco/src"],
+      "@webiny/api-aco-ddb/*": ["../api-aco-ddb/src/*"],
+      "@webiny/api-aco-ddb": ["../api-aco-ddb/src"],
+      "@webiny/api-audit-logs/*": ["../api-audit-logs/src/*"],
+      "@webiny/api-audit-logs": ["../api-audit-logs/src"],
+      "@webiny/api-audit-logs-ddb/*": ["../api-audit-logs-ddb/src/*"],
+      "@webiny/api-audit-logs-ddb": ["../api-audit-logs-ddb/src"],
+      "@webiny/api-core/*": ["../api-core/src/*"],
+      "@webiny/api-core": ["../api-core/src"],
+      "@webiny/api-core-ddb/*": ["../api-core-ddb/src/*"],
+      "@webiny/api-core-ddb": ["../api-core-ddb/src"],
+      "@webiny/api-file-manager/*": ["../api-file-manager/src/*"],
+      "@webiny/api-file-manager": ["../api-file-manager/src"],
+      "@webiny/api-file-manager-aco/*": ["../api-file-manager-aco/src/*"],
+      "@webiny/api-file-manager-aco": ["../api-file-manager-aco/src"],
+      "@webiny/api-file-manager-s3/*": ["../api-file-manager-s3/src/*"],
+      "@webiny/api-file-manager-s3": ["../api-file-manager-s3/src"],
+      "@webiny/api-headless-cms/*": ["../api-headless-cms/src/*"],
+      "@webiny/api-headless-cms": ["../api-headless-cms/src"],
+      "@webiny/api-headless-cms-aco/*": ["../api-headless-cms-aco/src/*"],
+      "@webiny/api-headless-cms-aco": ["../api-headless-cms-aco/src"],
+      "@webiny/api-headless-cms-ddb/*": ["../api-headless-cms-ddb/src/*"],
+      "@webiny/api-headless-cms-ddb": ["../api-headless-cms-ddb/src"],
+      "@webiny/api-headless-cms-scheduler/*": ["../api-headless-cms-scheduler/src/*"],
+      "@webiny/api-headless-cms-scheduler": ["../api-headless-cms-scheduler/src"],
+      "@webiny/api-headless-cms-tasks/*": ["../api-headless-cms-tasks/src/*"],
+      "@webiny/api-headless-cms-tasks": ["../api-headless-cms-tasks/src"],
+      "@webiny/api-headless-cms-workflows/*": ["../api-headless-cms-workflows/src/*"],
+      "@webiny/api-headless-cms-workflows": ["../api-headless-cms-workflows/src"],
+      "@webiny/api-mailer/*": ["../api-mailer/src/*"],
+      "@webiny/api-mailer": ["../api-mailer/src"],
+      "@webiny/api-record-locking/*": ["../api-record-locking/src/*"],
+      "@webiny/api-record-locking": ["../api-record-locking/src"],
+      "@webiny/api-scheduler/*": ["../api-scheduler/src/*"],
+      "@webiny/api-scheduler": ["../api-scheduler/src"],
+      "@webiny/api-scheduler-aws/*": ["../api-scheduler-aws/src/*"],
+      "@webiny/api-scheduler-aws": ["../api-scheduler-aws/src"],
+      "@webiny/api-website-builder/*": ["../api-website-builder/src/*"],
+      "@webiny/api-website-builder": ["../api-website-builder/src"],
+      "@webiny/api-website-builder-scheduler/*": ["../api-website-builder-scheduler/src/*"],
+      "@webiny/api-website-builder-scheduler": ["../api-website-builder-scheduler/src"],
+      "@webiny/api-website-builder-workflows/*": ["../api-website-builder-workflows/src/*"],
+      "@webiny/api-website-builder-workflows": ["../api-website-builder-workflows/src"],
+      "@webiny/api-websockets/*": ["../api-websockets/src/*"],
+      "@webiny/api-websockets": ["../api-websockets/src"],
+      "@webiny/api-websockets-aws/*": ["../api-websockets-aws/src/*"],
+      "@webiny/api-websockets-aws": ["../api-websockets-aws/src"],
+      "@webiny/api-websockets-ddb/*": ["../api-websockets-ddb/src/*"],
+      "@webiny/api-websockets-ddb": ["../api-websockets-ddb/src"],
+      "@webiny/api-workflows/*": ["../api-workflows/src/*"],
+      "@webiny/api-workflows": ["../api-workflows/src"],
+      "@webiny/aws-sdk/*": ["../aws-sdk/src/*"],
+      "@webiny/aws-sdk": ["../aws-sdk/src"],
+      "@webiny/background-tasks/*": ["../background-tasks/src/*"],
+      "@webiny/background-tasks": ["../background-tasks/src"],
+      "@webiny/cognito/*": ["../cognito/src/*"],
+      "@webiny/cognito": ["../cognito/src"],
+      "@webiny/event-handler-aws/*": ["../event-handler-aws/src/*"],
+      "@webiny/event-handler-aws": ["../event-handler-aws/src"],
+      "@webiny/handler/*": ["../handler/src/*"],
+      "@webiny/handler": ["../handler/src"],
+      "@webiny/handler-db/*": ["../handler-db/src/*"],
+      "@webiny/handler-db": ["../handler-db/src"],
+      "@webiny/handler-graphql/*": ["../handler-graphql/src/*"],
+      "@webiny/handler-graphql": ["../handler-graphql/src"],
+      "@webiny/webhooks/*": ["../webhooks/src/*"],
+      "@webiny/webhooks": ["../webhooks/src"]
+    }
+  }
+}
