+++ packages/api-infra-aws/tsconfig.json
@@ -1,0 +1,117 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src", "__tests__"],
+  "references": [
+    { "path": "../api-aco" },
+    { "path": "../api-aco-ddb" },
+    { "path": "../api-audit-logs" },
+    { "path": "../api-audit-logs-ddb" },
+    { "path": "../api-core" },
+    { "path": "../api-core-ddb" },
+    { "path": "../api-file-manager" },
+    { "path": "../api-file-manager-aco" },
+    { "path": "../api-file-manager-s3" },
+    { "path": "../api-headless-cms" },
+    { "path": "../api-headless-cms-aco" },
+    { "path": "../api-headless-cms-ddb" },
+    { "path": "../api-headless-cms-scheduler" },
+    { "path": "../api-headless-cms-tasks" },
+    { "path": "../api-headless-cms-workflows" },
+    { "path": "../api-mailer" },
+    { "path": "../api-record-locking" },
+    { "path": "../api-scheduler" },
+    { "path": "../api-scheduler-aws" },
+    { "path": "../api-website-builder" },
+    { "path": "../api-website-builder-scheduler" },
+    { "path": "../api-website-builder-workflows" },
+    { "path": "../api-websockets" },
+    { "path": "../api-websockets-aws" },
+    { "path": "../api-websockets-ddb" },
+    { "path": "../api-workflows" },
+    { "path": "../aws-sdk" },
+    { "path": "../background-tasks" },
+    { "path": "../cognito" },
+    { "path": "../event-handler-aws" },
+    { "path": "../handler" },
+    { "path": "../handler-db" },
+    { "path": "../handler-graphql" },
+    { "path": "../webhooks" }
+  ],
+  "compilerOptions": {
+    "rootDirs": ["./src", "./__tests__"],
+    "outDir": "./dist",
+    "declarationDir": "./dist",
+    "paths": {
+      "~/*": ["./src/*"],
+      "~tests/*": ["./__tests__/*"],
+      "@webiny/api-aco/*": ["../api-aco/src/*"],
+      "@webiny/api-aco": ["../api-aco/src"],
+      "@webiny/api-aco-ddb/*": ["../api-aco-ddb/src/*"],
+      "@webiny/api-aco-ddb": ["../api-aco-ddb/src"],
+      "@webiny/api-audit-logs/*": ["../api-audit-logs/src/*"],
+      "@webiny/api-audit-logs": ["../api-audit-logs/src"],
+      "@webiny/api-audit-logs-ddb/*": ["../api-audit-logs-ddb/src/*"],
+      "@webiny/api-audit-logs-ddb": ["../api-audit-logs-ddb/src"],
+      "@webiny/api-core/*": ["../api-core/src/*"],
+      "@webiny/api-core": ["../api-core/src"],
+      "@webiny/api-core-ddb/*": ["../api-core-ddb/src/*"],
+      "@webiny/api-core-ddb": ["../api-core-ddb/src"],
+      "@webiny/api-file-manager/*": ["../api-file-manager/src/*"],
+      "@webiny/api-file-manager": ["../api-file-manager/src"],
+      "@webiny/api-file-manager-aco/*": ["../api-file-manager-aco/src/*"],
+      "@webiny/api-file-manager-aco": ["../api-file-manager-aco/src"],
+      "@webiny/api-file-manager-s3/*": ["../api-file-manager-s3/src/*"],
+      "@webiny/api-file-manager-s3": ["../api-file-manager-s3/src"],
+      "@webiny/api-headless-cms/*": ["../api-headless-cms/src/*"],
+      "@webiny/api-headless-cms": ["../api-headless-cms/src"],
+      "@webiny/api-headless-cms-aco/*": ["../api-headless-cms-aco/src/*"],
+      "@webiny/api-headless-cms-aco": ["../api-headless-cms-aco/src"],
+      "@webiny/api-headless-cms-ddb/*": ["../api-headless-cms-ddb/src/*"],
+      "@webiny/api-headless-cms-ddb": ["../api-headless-cms-ddb/src"],
+      "@webiny/api-headless-cms-scheduler/*": ["../api-headless-cms-scheduler/src/*"],
+      "@webiny/api-headless-cms-scheduler": ["../api-headless-cms-scheduler/src"],
+      "@webiny/api-headless-cms-tasks/*": ["../api-headless-cms-tasks/src/*"],
+      "@webiny/api-headless-cms-tasks": ["../api-headless-cms-tasks/src"],
+      "@webiny/api-headless-cms-workflows/*": ["../api-headless-cms-workflows/src/*"],
+      "@webiny/api-headless-cms-workflows": ["../api-headless-cms-workflows/src"],
+      "@webiny/api-mailer/*": ["../api-mailer/src/*"],
+      "@webiny/api-mailer": ["../api-mailer/src"],
+      "@webiny/api-record-locking/*": ["../api-record-locking/src/*"],
+      "@webiny/api-record-locking": ["../api-record-locking/src"],
+      "@webiny/api-scheduler/*": ["../api-scheduler/src/*"],
+      "@webiny/api-scheduler": ["../api-scheduler/src"],
+      "@webiny/api-scheduler-aws/*": ["../api-scheduler-aws/src/*"],
+      "@webiny/api-scheduler-aws": ["../api-scheduler-aws/src"],
+      "@webiny/api-website-builder/*": ["../api-website-builder/src/*"],
+      "@webiny/api-website-builder": ["../api-website-builder/src"],
+      "@webiny/api-website-builder-scheduler/*": ["../api-website-builder-scheduler/src/*"],
+      "@webiny/api-website-builder-scheduler": ["../api-website-builder-scheduler/src"],
+      "@webiny/api-website-builder-workflows/*": ["../api-website-builder-workflows/src/*"],
+      "@webiny/api-website-builder-workflows": ["../api-website-builder-workflows/src"],
+      "@webiny/api-websockets/*": ["../api-websockets/src/*"],
+      "@webiny/api-websockets": ["../api-websockets/src"],
+      "@webiny/api-websockets-aws/*": ["../api-websockets-aws/src/*"],
+      "@webiny/api-websockets-aws": ["../api-websockets-aws/src"],
+      "@webiny/api-websockets-ddb/*": ["../api-websockets-ddb/src/*"],
+      "@webiny/api-websockets-ddb": ["../api-websockets-ddb/src"],
+      "@webiny/api-workflows/*": ["../api-workflows/src/*"],
+      "@webiny/api-workflows": ["../api-workflows/src"],
+      "@webiny/aws-sdk/*": ["../aws-sdk/src/*"],
+      "@webiny/aws-sdk": ["../aws-sdk/src"],
+      "@webiny/background-tasks/*": ["../background-tasks/src/*"],
+      "@webiny/background-tasks": ["../background-tasks/src"],
+      "@webiny/cognito/*": ["../cognito/src/*"],
+      "@webiny/cognito": ["../cognito/src"],
+      "@webiny/event-handler-aws/*": ["../event-handler-aws/src/*"],
+      "@webiny/event-handler-aws": ["../event-handler-aws/src"],
+      "@webiny/handler/*": ["../handler/src/*"],
+      "@webiny/handler": ["../handler/src"],
+      "@webiny/handler-db/*": ["../handler-db/src/*"],
+      "@webiny/handler-db": ["../handler-db/src"],
+      "@webiny/handler-graphql/*": ["../handler-graphql/src/*"],
+      "@webiny/handler-graphql": ["../handler-graphql/src"],
+      "@webiny/webhooks/*": ["../webhooks/src/*"],
+      "@webiny/webhooks": ["../webhooks/src"]
+    }
+  }
+}
